+++ src/utils/getFieldsWithRelationshipOptions.js
@@ -1,11 +1,27 @@
 import setOptions from './setOptions.js'
 
-export default async function ({ fields, idKey, parent, relationships }) {
+export default async function ({ fields, idKey, parent, relationships, form, id }) {
   fields = JSON.parse(JSON.stringify(fields))
+
+  const methodsModels = {
+    fetchList: !form && !id,
+    fetchSingleCreate: form && !id,
+    fetchSingleEdit: form && id,
+    fetchSingleShow: !form && id
+  }
+
+  // model with value true
+  const model = Object.keys(methodsModels).find(item => methodsModels[item])
 
   for (const key in relationships) {
     const relationship = relationships[key]
-    const { ref } = relationship
+    const { ref, methods } = relationship
+
+    // if there are methods, and it is not the current method returns empty options
+    if (methods && methods.length && !methods.includes(model)) {
+      fields[key].options = []
+      continue
+    }
 
     if (ref) {
       if (!relationship.label) { relationship.label = idKey }