+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+### Added
+- added option to send relationship in filters.
+- added `preQueryList` function to intercept queries.
+- added the possibility of having relationship of collections only for declared methods.
+
+### Changed
+- changed base url in `makesync` function.
+
+### Fixed
+- calculate progress only when pulling data in sync.
+
 # 2.2.1 2021-07-16
 
 ### Added