+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 2.7.0 - 2021-10-14
+
+### Added
+- added function `onSync` in constructor to get sync.
+
+### Removed
+- removed function `progress` in constructor.
+
 # 2.6.0 - 2021-09-07
 
 ### Added